requirements.txt
flask
beautifulsoup4
requests
dnspython
python-whois
